+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 1.9.1 (not released yet)
 
+* Added `DataType.registeredType`.
+* Added support for `RecordType` inheritance.
 * Added support for automatic use of auth tokens.
 
 ## 1.9.0 (2019-04-15)